+++ src/app/content/App/Help/Help.stories.tsx
@@ -6,10 +6,10 @@ import { action } from '@storybook/addon-actions';
 import Notification from 'components/organisms/Notification';
 import { Help } from '.';
 
-storiesOf('screens/Help', module)
+storiesOf('Extension', module)
   .addDecorator(getStory => (
     <Router>
       <Notification close={action('close')}>{getStory()}</Notification>
     </Router>
   ))
-  .add('normal', () => <Help />);
+  .add('Help', () => <Help />);