+++ src/app/content/App/Question/FormScreen/FormScreen.stories.tsx
@@ -5,7 +5,7 @@ import Notification from 'components/organisms/Notification';
 import FormScreen from './';
 import { formStoreDecorator } from '../../../../../../.storybook/decorators';
 
-storiesOf('screens/Question/Form', module)
+storiesOf('Extension/Question/Form', module)
   .addDecorator(formStoreDecorator)
   .addDecorator(getStory => (
     <Router>