+++ src/components/atoms/Button/BackgroundButton/BackgroundButton.stories.tsx
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import { withKnobs, text } from '@storybook/addon-knobs';
 import BackgroundButton from './BackgroundButton';
 
-storiesOf('atoms/Buttons/BackgroundButton', module)
+storiesOf('Components/Atoms/Buttons/BackgroundButton', module)
   .addDecorator(withKnobs)
   .add('normal', () => (
     <BackgroundButton