+++ src/components/atoms/Button/OpenButton/OpenButton.stories.tsx
@@ -2,6 +2,6 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import OpenButton from './OpenButton';
 
-storiesOf('atoms/Buttons/OpenButton', module).add('normal', () => (
+storiesOf('Components/Atoms/Buttons', module).add('OpenButton', () => (
   <OpenButton />
 ));