+++ src/components/atoms/Forms/Error/Error.stories.tsx
@@ -2,6 +2,6 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import Error from './Error';
 
-storiesOf('atoms/Forms/Error Text', module).add('text', () => (
+storiesOf('Components/Atoms/Forms/Error Text', module).add('text', () => (
   <Error>Les champs en rouge sont incorrects ou incomplets</Error>
 ));