+++ src/components/atoms/Tab/Tab.stories.tsx
@@ -2,6 +2,6 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import Tab from './Tab';
 
-storiesOf('atoms/Tab', module)
+storiesOf('Components/Atoms/Tab', module)
   .add('normal', () => <Tab>Mes abonnements</Tab>)
   .add('active', () => <Tab active>Mes abonnements</Tab>);