+++ src/components/molecules/TabsNav/TabsNav.stories.tsx
@@ -2,4 +2,6 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import TabsNav from './TabsNav';
 
-storiesOf('molecules/TabNav', module).add('normal', () => <TabsNav />);
+storiesOf('Components/Molecules/TabNav', module).add('normal', () => (
+  <TabsNav />
+));